requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langgraph>=0.2.0          
langchain-groq>=0.2.0   
sqlalchemy[asyncio]==2.0.27
aiopg==0.16.0
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
